requirements.txt
# Python dependencies for Swing Trading Bot

# Interactive Brokers API
ib_insync>=0.9.86

# Configuration
PyYAML>=6.0

# Data handling (optional but recommended)
pandas>=2.0.0
numpy>=1.24.0

# Note: Install with: pip install -r requirements.txt
